requirements.txt
discord.py
pycryptodome
binascii
asyncio
dotenv